Always use the same window to display
This offers almost the same experience as the cscope plugin in vim.

It's really hard to manage multiple windows when emacs is spliting new
window or overwriting another window automatically. If I need a new
window, I will type '<C-w>v' or ': vs" quickly in evil mode. It feels
natrual and it is much easier to get the expected result this way.